- add TheCalculatorEngine - CalculatorEngine and RPNCalculatorEngine extends TheCalculatorEngine - move some method to TheCalculatorEngine
